+++ lib/option_result.dart
@@ -1,8 +1,21 @@
+/// This is the base `option_result` library, containing both [Option] and [Result]
+/// types and their helpers.
+///
+/// This library can be imported via:
+///
+/// ```dart
+/// import 'package:option_result/option_result.dart';
+/// ```
+///
+/// If you want to only import the [Option] or the [Result] types individually,
+/// do so via:
+///
+/// ```dart
+/// import 'package:option_result/option.dart';
+/// // or
+/// import 'package:option_result/result.dart';
+/// ```
 library option_result;
 
-import 'dart:async';
-
-part 'src/helpers.dart';
-part 'src/option.dart';
-part 'src/result.dart';
-part 'src/util.dart';
+export 'option.dart';
+export 'result.dart';