+++ lib/src/util.dart
@@ -1,4 +1,2 @@
-part of option_result;
-
 /// Compares the runtime types of the given objects.
-bool _compareRuntimeTypes(Object a, Object b) => a.runtimeType.hashCode == b.runtimeType.hashCode;
+bool compareRuntimeTypes(Object a, Object b) => a.runtimeType == b.runtimeType;