+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.1-dev-2
+
+- Add separate packages to allow importing option and result separately
+- Add `Option#filter()` method
+
 ## 0.0.1-dev-1
 
 - Initial version.